add parse command, optimize convert command

+++ scripts/audio_cli.js
@@ -40,7 +40,23 @@ yargs
     "Test command",
     (yargs) => {},
     (argv) => {
-      readTagsFromDatabase(argv._[0]);
+      cmdCheckFiles(argv._[1]);
+    }
+  )
+  .command(
+    "clean <source>",
+    "Clean invalid audio files in source dir",
+    (yargs) => {},
+    (argv) => {
+      cmdCleanInvalid(argv.source);
+    }
+  )
+  .command(
+    ["parse <source> [options]", "rd"],
+    "Parse exif for audio files and save to database",
+    (yargs) => {},
+    (argv) => {
+      cmdParseTags(argv.source);
     }
   )
   .command(
@@ -59,11 +75,11 @@ yargs
         .option("force", {
           alias: "f",
           type: "boolean",
-          description: "Force to override exists file",
+          describe: "Force to override exists file",
         });
     },
     (argv) => {
-      commandConvert(argv);
+      cmdConvert(argv);
     }
   )
   .command(
@@ -76,7 +92,7 @@ yargs
       });
     },
     (argv) => {
-      commandCheckTags(argv);
+      cmdCheckFiles(argv);
     }
   )
   .command(
@@ -92,21 +108,20 @@ yargs
           alias: "l",
           type: "array",
           default: [],
-          description: "Audio language that should be moved (cn,ja,kr,en)",
+          describe: "Audio language that should be moved (cn,ja,kr,en)",
         })
         .option("ignore", {
           alias: "g",
           type: boolean,
-
-          description: "Ingore rare language audio files (don't move)",
+          describe: "Ingore rare language audio files (don't move)",
         });
     },
     (argv) => {
-      commandMoveByLng(argv);
+      cmdMoveByLng(argv);
     }
   ).argv;
 
-async function commandConvert(argv) {
+async function cmdConvert(argv) {
   const root = path.resolve(argv.source);
   if (!root || !fs.pathExistsSync(root)) {
     yargs.showHelp();
@@ -120,31 +135,37 @@ async function checkFiles(files) {
   const results = await Promise.all(
     // true means keep
     // false mean skip
-    files.map(async (f, i) => {
+    files.map(async (file, i) => {
+      const f = file.path;
       const index = i + 1;
+      if (!(await fs.pathExists(f))) {
+        d.I(`SkipNotFound (${index}) ${h.ps(f)}`);
+        return false;
+      }
       if (h.ext(f, true) == ".m4a") {
-        d.L(chalk.gray(`SkipAAC (${index}): ${h.ps(f)}`));
+        d.D(chalk.gray(`SkipAAC (${index}): ${h.ps(f)}`));
         return false;
       }
       const aacName = h.getAACFileName(f);
       const p1 = path.join(path.dirname(f), "output", aacName);
       if (await fs.pathExists(p1)) {
-        d.W(chalk.gray(`SkipExists (${i}): ${h.ps(p1)}`));
+        d.D(chalk.gray(`SkipExists (${i}): ${h.ps(p1)}`));
         return false;
       }
       const p2 = path.join(path.dirname(f), aacName);
       if (await fs.pathExists(p2)) {
-        d.W(chalk.gray(`SkipExists (${index}): ${h.ps(p2)}`));
+        d.D(chalk.gray(`SkipExists (${index}): ${h.ps(p2)}`));
         return false;
       }
-      d.L(chalk.green(`Prepared (${index}): `) + `${h.ps(f)}`);
+      d.I(chalk.green(`Prepared (${index}): `) + `${h.ps(f)}`);
       return true;
     })
   );
   return files.filter((_v, i) => results[i]);
 }
 
-async function allToAAC(files) {
+async function convertAllToAAC(files) {
+  d.L(`convertAllToAAC: Adding ${files.length} tasks`);
   const pool = workerpool.pool(__dirname + "/audio_workers.js", {
     maxWorkers: cpuCount,
     workerType: "process",
@@ -152,33 +173,73 @@ async function allToAAC(files) {
   const startMs = Date.now();
   const results = await Promise.all(
     files.map(async (f, i) => {
-      const file = path.resolve(f);
-      const result = await pool.exec("toAAC", [file, i + 1]);
+      const result = await pool.exec("toAAC", [f, i + 1]);
       return result;
     })
   );
   await pool.terminate();
-  const elapsedSecs = (Date.now() - startMs) / 1000;
-  d.L(`Result: ${results.length} files converted in ${elapsedSecs}s.`);
+  d.L(`Result: ${results.length} files converted in ${h.ht(startMs)}.`);
   return results;
 }
 
+function appendAudioBitRate(f) {
+  const bitRateTag = f.tags && f.tags.AudioBitrate;
+  if (!bitRateTag) {
+    return f;
+  }
+  const r = /(\d+)\.?\d*/;
+  let bitRate = parseInt(bitRateTag);
+  if (!bitRate) {
+    const m = r.exec(bitRateTag);
+    bitRate = parseInt(m && m[0]) || 0;
+  }
+  if (bitRate < 192) {
+    d.D(
+      `appendAudioBitRate: ${bitRate} ${path.basename(f.path)} ${
+        f.tags.MIMEType
+      } ${f.tags.AudioBitrate}`
+    );
+  }
+  f.bitRate = bitRate;
+  return f;
+}
+
 async function executeConvert(root) {
+  d.L(`executeConvert: ${root}`);
   const startMs = Date.now();
   // list all files in dir recursilly
-  let files = klawSync(root, { nodir: true }).map((f) => f.path);
+  const taskFiles = klawSync(root, { nodir: true });
+  const taskPaths = taskFiles.map((f) => f.path);
+  d.L(`Total ${taskPaths.length} files found in ${h.ht(startMs)}`);
+  // caution: slow on network drives
+  // files = await exif.readAllTags(files);
+  // files = files.filter((f) => h.isAudioFile(f.path));
+  // saveAudioDBTags(files);
+  // use cached file with tags database
+  let files = await readAudioDBTags(root);
+  d.L(`Total ${files.length} files parsed in ${h.ht(startMs)}`);
+  files = files.filter((f) => taskPaths.includes(f.path));
+  if (files.length == 0) {
+    // new files not found in db
+    // parse exif tags and save to db
+    files = await exif.readAllTags(taskFiles);
+    console.log(files.length);
+    files = files.filter((f) => h.isAudioFile(f.path));
+    await saveAudioDBTags(files);
+  }
+  files = files.map((f) => appendAudioBitRate(f));
+  d.L(`Total ${files.length} files after filterd in ${h.ht(startMs)}`);
   const filesCount = files.length;
   // keep only non-m4a audio files
   // todo add check to ensure is audio file
   files = await checkFiles(files);
-  d.L(`Total ${filesCount} files processed in ${Date.now() - startMs}ms`);
+
+  d.L(`Total ${filesCount} files checked in ${h.ht(startMs)}`);
   const skipCount = filesCount - files.length;
   if (skipCount > 0) {
     d.L(`Total ${skipCount} audio files are skipped`);
   }
-  const output = path.join(root, "output");
   d.L(`Input: ${root}`);
-  d.L(`Output: ${output}`);
   if (files.length == 0) {
     d.L(chalk.green("Nothing to do, exit now."));
     return;
@@ -193,9 +254,8 @@ async function executeConvert(root) {
     },
   ]);
   if (answer.yes) {
-    allToAAC(files).then((results) => {
-      d.L(chalk.green(`There are ${results.length} audio files converted.`));
-    });
+    const results = await convertAllToAAC(files);
+    d.L(chalk.green(`There are ${results.length} audio files converted.`));
   } else {
     d.L(chalk.yellowBright("Will do nothing, aborted by user."));
   }
@@ -211,49 +271,19 @@ async function commandCheckTags(argv) {
   await executeCheckTags(root);
 }
 
-async function updateID3Tag(files) {
+async function updateID3TagForFile(f) {
   // https://id3.org/id3v2.3.0
   // https://www.npmjs.com/package/node-id3
-  for (const f of files) {
-    const tags = { title: f.tags.Title, artist: f.tags.Artist };
-    const ret = id3.update(tags, f);
+  const tags = { title: f.tags.Title, artist: f.tags.Artist };
+  try {
+    const ret = await id3.update(tags, f.path);
+    d.L(`Update ID3 OK for ${f.path}`);
+  } catch (error) {
+    d.E(`Update ID3 Error:${error} for ${f.path}`);
   }
 }
 
-async function readTagsFromFile(root) {
-  const jsonName = sanitize(root, { replacement: "_" });
-  const jsonPath = path.join("data", `alltags.json`);
-  const startMs = Date.now();
-  const data = await fs.readFile(jsonPath, { encoding: "utf8" });
-  const lines = data
-    .split("\n")
-    .filter((l) => typeof l === "string" && l.startsWith("{"));
-  d.L(`readTagsFromFile: ${lines.length} lines from ${jsonPath}`);
-  const tagMap = new Map();
-  await Promise.all(
-    lines.map(async (line, i) => {
-      try {
-        const tags = JSON.parse(line);
-        const key = h.replaceAll(tags.SourceFile, "/", path.sep);
-        tagMap.set(key, tags);
-      } catch (error) {
-        d.E(error);
-      }
-    })
-  );
-  // const tagMap = tagArr.reduce((map, obj) => {
-  //   map[obj.path] = obj.tags;
-  // }, {});
-  d.L(`readTagsFromFile: ${tagMap.size} tags has loaded (${h.ht(startMs)})`);
-  return tagMap;
-}
-
-async function saveTagsToDatabase(files) {
-  sqlite3.verbose();
-  const db = await sqlite.open({
-    filename: "./data/audio.db",
-    driver: sqlite3.Database,
-  });
+async function createAudioTable(db) {
   // https://www.npmjs.com/package/sqlite
   await db.exec(
     `CREATE TABLE IF NOT EXISTS tags (
@@ -267,69 +297,220 @@ async function saveTagsToDatabase(files) {
       PRIMARY KEY(path)
     );`
   );
+  return db;
+}
+
+async function openAudioDB(dbFile) {
+  sqlite3.verbose();
+  const db = await sqlite.open({
+    filename: dbFile || "./data/audio.db",
+    driver: sqlite3.Database,
+  });
+  await createAudioTable(db);
+  return db;
+}
+
+async function saveAudioDBRow(db, f) {
+  if (!(db && f)) {
+    throw new Error("Database and file object is required!");
+  }
+  const ret = await db.run(
+    "INSERT OR REPLACE INTO tags VALUES (?,?,?,?,?,?)",
+    f.size || f.stats.size || 0,
+    path.basename(f.path),
+    f.path,
+    f.tags.Title || "",
+    f.tags.Artist || "",
+    JSON.stringify(f.tags)
+  );
+  d.D(`saveAudioDBRow: row ${ret.lastID} for ${f.path} `);
+}
+
+async function saveAudioDBTags(files) {
+  // const dbFile = "./data/audio.db";
+  // if (await fs.pathExists(dbFile)) {
+  //   await fs.move(dbFile, dbFile + "." + Date.now());
+  // }
+  const db = await openAudioDB();
   const dbStartMs = Date.now();
   db.run("BEGIN TRANSACTION");
   try {
     for (const f of files) {
-      const ret = await db.run(
-        "INSERT OR REPLACE INTO tags VALUES (?,?,?,?,?,?)",
-        f.stats.size,
-        path.basename(f.path),
-        f.path,
-        f.tags.Title,
-        f.tags.Artist,
-        JSON.stringify(f.tags)
-      );
-      d.I(`Insert row ${ret.lastID} for ${f.path} `);
+      await saveAudioDBRow(db, f);
     }
   } catch (error) {
     d.E(error);
   }
   db.run("COMMIT TRANSACTION");
   await db.close();
-  d.L(`Database insert ${files.length} rows in  ${h.ht(dbStartMs)}`);
+  d.L(`saveAudioDBTags: ${files.length} rows added in ${h.ht(dbStartMs)}`);
 }
 
-async function readTagsFromDatabase(root) {
+async function readAudioDBTags(root) {
   const dbStartMs = Date.now();
-  sqlite3.verbose();
-  const db = await sqlite.open({
-    filename: "./data/audio.db",
-    driver: sqlite3.Database,
-  });
-  const rows = await db.all("SELECT path,tags FROM tags");
-  const tagMap = new Map();
-  await Promise.all(
+  const db = await openAudioDB();
+  const rows = await db.all("SELECT * FROM tags");
+  const files = await Promise.all(
     rows.map(async (row, i) => {
       try {
-        const key = row.path;
-        const tags = JSON.parse(row.tags);
-        tagMap.set(key, tags);
-        d.I(`Read row ${i} for ${key}`);
+        d.D(`Read row ${i} ${row.path} ${row.size}`);
+        return {
+          path: row.path,
+          size: row.size,
+          tags: JSON.parse(row.tags),
+        };
       } catch (error) {
         d.E(error);
       }
     })
   );
+  await db.close();
   d.L(`Database read ${rows.length} rows in ${h.ht(dbStartMs)}`);
-  return tagMap;
+  return files;
+}
+
+async function cmdParseTags(root) {
+  if (!root || !fs.pathExistsSync(root)) {
+    yargs.showHelp();
+    d.E(chalk.red(`ERROR! Source '${root}' is not exists or not a directory!`));
+    return;
+  }
+  await executeParseTags(root);
+}
+
+async function executeParseTags(root) {
+  // miragate json file to sqlite db
+  // try {
+  //   let files = await fs.readJSON("./data/alltags.json");
+  //   files = files.filter((f) => f.tags && f.tags.Artist && f.tags.Title);
+  //   await saveAudioDBTags(files);
+  // } catch (error) {
+  //   d.E(error);
+  // }
+
+  d.L(`Input: ${root}`);
+  let startMs = Date.now();
+  let files = exif.listFiles(root, (f) => h.isAudioFile(f.path));
+  const fileCount = files.length;
+  d.L(`executeParseTags: ${fileCount} files found in (${h.ht(startMs)})`);
+  startMs = Date.now();
+  // two slow over network
+  files = await exif.readAllTags(files);
+  d.L(`executeParseTags: ${fileCount} files parsed in ${h.ht(startMs)}`);
+  try {
+    const jsonName = sanitize(root);
+    await fs.writeJSON(`./data/${jsonName}.json`, files);
+    d.L(`executeParseTags: JSON ./data/${jsonName}.json`);
+  } catch (error) {
+    d.E("executeParseTags:", error);
+  }
+  await saveAudioDBTags(files);
+}
+
+async function cmdCheckFiles(root) {
+  d.L(`Input: ${root}`);
+  const startMs = Date.now();
+  let files = exif.listFiles(root, (f) => h.isAudioFile(f.path));
+  files = exif.readAllTags(files);
+  for (const f of files) {
+    const name = path.basename(f.path);
+    if (/\d+/.test(name)) {
+      d.L(`check: ${h.ps(f.path)}`);
+    }
+  }
+}
+
+async function cmdCleanInvalid(root) {
+  d.L(`Input: ${root}`);
+  const startMs = Date.now();
+  let files = exif.listFiles(root, (f) => h.isAudioFile(f.path));
+  for (const f of files) {
+    const name = path.basename(f.path);
+    if (!name.includes(" @ ")) {
+      d.L(`Delete Invalid: ${h.ps(f.path)}`);
+      await fs.rm(f.path);
+    }
+  }
+}
+
+async function fixAllAudioTags() {
+  const files = await readAudioDBTags();
+  const db = await openAudioDB();
+  let changedCount = 0;
+  for (const f of files) {
+    const filename = f.path;
+    // d.L(`Processing: ${filename}`);
+    const tags = f.tags;
+    const title = tags.Title;
+    const artist = tags.Artist;
+    const reSep = /[/&(:（：]/;
+    let newTitle = null;
+    let newArtist = null;
+    if (un.strOnlyNonWordChars(artist) || un.strOnlyNonWordChars(title)) {
+      // 删除所有标签乱码的文件，并从数据库移除
+      try {
+        if (await fs.pathExists(filename)) {
+          await fs.rm(filename);
+          d.L(`Deleted ${artist} - ${title}  => ${h.ps(filename)}`);
+        }
+        const ret = await db.run("DELETE FROM tags WHERE path = ?", filename);
+        d.L(`Remove from DB: ${ret.changes} ${filename}`);
+      } catch (error) {
+        d.E(error);
+      }
+    } else if (un.strHasNonWordChars(artist)) {
+      if (reSep.test(artist)) {
+        let n = null;
+        const parts = artist.split(reSep);
+        if (parts.length > 1) {
+          n = parts[0] || parts[1];
+        } else {
+          n = parts[0];
+        }
+        n = n.trim();
+        if (n != artist) {
+          newArtist = n;
+          d.L(
+            `New Artist: '${artist}' to '${newArtist}' for ${h.ps(filename)}`
+          );
+        }
+      }
+    } else if (
+      filename.includes("CN") &&
+      un.strHasHanyu(artist) &&
+      artist.length > 4
+    ) {
+      d.L(`ArtistInvalid: ${artist} => ${h.ps(filename)}`);
+    } else {
+      // d.L(`Artist: ${artist} => `);
+    }
+    if (newArtist) {
+      newArtist && (f.tags.Artist = newArtist);
+      d.L(`Changed(${++changedCount}): ${f.tags.Artist} - ${f.tags.Title}`);
+      await updateID3TagForFile(f);
+      await saveTagsToDBForFile(db, f);
+    }
+  }
 }
 
 async function executeCheckTags(root) {
   d.L(`Input: ${root}`);
   const startMs = Date.now();
-  let files = exif.listFiles(root, (f) => h.isAudioFile(f.path));
-  const fileCount = files.length;
-  d.L(`executeCheckTags: total ${fileCount} files found in (${h.ht(startMs)})`);
+  // let files = exif.listFiles(root, (f) => h.isAudioFile(f.path));
+  // const fileCount = files.length;
+  // d.L(`executeCheckTags: total ${fileCount} files found in (${h.ht(startMs)})`);
+
   // const tagMap = await readTagsFromFile(root);
-  const tagMap = await readTagsFromDatabase(root);
-  files = files.map((f) => {
-    f.tags = tagMap.get(f.path);
-    return f;
-  });
-  // files = await exif.readTags(files);
-  files = files.filter((f) => f.tags);
+  // const tagMap = await readTagsFromDatabaseAsMap(root);
+  // files = files.map((f) => {
+  //   f.tags = tagMap.get(f.path);
+  //   return f;
+  // });
+  // two slow over network
+  // files = await exif.readAllTags(files);
   // saveTagsToDatabase(files);
+  let files = await readTagsFromDatabase(root);
+  const fileCount = files.length;
   if (files.length < fileCount) {
     d.W(
       chalk.yellow(
@@ -343,42 +524,20 @@ async function executeCheckTags(root) {
     )})`
   );
 
-  // (await fs.pathExists(jsonPath)) &&
-  //   (await fs.move(jsonPath, jsonPath + `.${Date.now()}`));
-  // let i = 0;
-  // for (const f of files) {
-  //   await fs.appendFile(jsonPath, JSON.stringify(f.tags) + "\n");
-  //   if (++i % 1000 == 0) {
-  //     d.L(`WriteJson[${++i}/${files.length}]: ${h.ht(startMs)}`);
-  //   }
-  // }
+  for (const f of files) {
+    const name = path.basename(f.path, path.extname(f.path));
+    if (/[~!#\$%\^&\*_+-={}]/.test(name)) {
+      d.L(`Checked ${f.path} ${f.tags.Artist} @ ${f.tags.Title}`);
+    } else {
+      // d.L(`Skip ${h.ps(f.path)}`);
+    }
+  }
 
-  // const iout = path.join(path.dirname(root), "invalid");
-  // if (!fs.pathExistsSync(iout)) {
-  //   fs.mkdirSync(iout);
-  // }
-  // files.forEach((f, i) => {
-  //   const t = f.tags;
-  //   if (!t.Artist || !t.Title) {
-  //     d.L(`Move Invalid: ${f.path}`);
-  //     // fs.moveSync(f.path, path.join(iout, path.basename(f.path)));
-  //   }
-  // });
-  d.L(`Input: ${root}`);
   d.L(`Processed ${files.length} files in ${h.ht(startMs)}`);
-
-  const answer = await inquirer.prompt([
-    {
-      type: "confirm",
-      name: "yes",
-      default: false,
-      message: chalk.bold.red(`Are you sure to check these files?`),
-    },
-  ]);
 }
 
-async function commandMoveByLng(argv) {
-  d.I(`commandMoveByLng:`, argv);
+async function cmdMoveByLng(argv) {
+  d.I(`cmdMoveByLng:`, argv);
   const root = path.resolve(argv.source);
   const lng = argv.lng || [];
   if (!root || !fs.pathExistsSync(root)) {
@@ -410,10 +569,14 @@ async function executeMoveByLng(root, lng = []) {
   d.I(outputs);
   const startMs = Date.now();
   let files = exif.listFiles(root);
-  files = await exif.readTags(files);
+  files = files.filter((f) => h.isAudioFile(f.path));
+  d.L(`executeMoveByLng: files count`, files.length);
+  files = await exif.readAllTags(files);
   files = files.filter((f) => {
     return f.tags && f.tags.Title && f.tags.Artist;
   });
+  d.L(`executeMoveByLng: tags count`, files.length);
+  // let files = await readTagsFromDatabase(root);
   const fileCount = files.length;
   files.forEach((f, i) => {
     const t = f.tags;
@@ -491,17 +654,26 @@ async function executeMoveByLng(root, lng = []) {
       await fs.mkdir(dout);
     }
     async function ensureMove(src, dst) {
+      d.L(`ensureMove: ${h.ps(src)} => ${h.ps(dst)}`);
       if (src == dst) {
-        d.W(`Skip:${src}`);
+        // d.W(`Skip:${src}`);
         return;
       }
-      if (await fs.pathExists(dst)) {
-        d.W(`Duplicate:${src}`);
-        await fs.move(src, path.join(dout, path.basename(src)));
-      } else {
-        d.D(`Moving to ${dst}`);
-        await fs.move(src, dst);
-        d.I(`Moved to ${dst}`);
+      if (await fs.pathExists(src)) {
+        // d.W(`NotExists:${src}`);
+        return;
+      }
+      try {
+        if (await fs.pathExists(dst)) {
+          // d.W(`Duplicate:${src}`);
+          await fs.move(src, path.join(dout, path.basename(src)));
+        } else {
+          // d.D(`Moving to ${dst}`);
+          await fs.move(src, dst);
+          d.I(`Moved to ${dst}`);
+        }
+      } catch (error) {
+        d.E(`ensureMove:${error}`);
       }
     }
     // https://zellwk.com/blog/async-await-in-loops/
